Programming_Clean/projects/ai-data-chat/requirements.txt
streamlit==1.39.0
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.3
openpyxl==3.1.5
